Add rendering utility classes
This commit adds a few classes designed to hide Qt's RHI usage
away from the plot renderers, and to share common code between them.
In addition, a 'pipegen' tool is added which is used at build time
to generate C++ code based on a json file which references a set of
shaders and adds some more information. The generated code automates
the creation of the pipelines and enforces type safety on the data
exchanged between the CPU and the GPU.